+++ src/Parser.php
@@ -1,6 +1,7 @@
 <?php
 namespace VovanVE\parser;
 
+use VovanVE\parser\actions\ActionAbortException;
 use VovanVE\parser\actions\ActionsMap;
 use VovanVE\parser\common\BaseObject;
 use VovanVE\parser\common\InternalException;
@@ -153,6 +154,11 @@ class Parser extends BaseObject
                 NEXT_SYMBOL:
             }
             DONE:
+        } catch (ActionAbortException $e) {
+            throw new SyntaxException(
+                $e->getMessage(),
+                $token ? $token->getOffset() : strlen($input)
+            );
         } catch (NoReduceException $e) {
             // This unexpected reduce (no rule to reduce) may happen only
             // when current terminal is not expected. So, some terminals