+++ src/actions/ActionAbortException.php
@@ -1,0 +1,6 @@
+<?php
+namespace VovanVE\parser\actions;
+
+class ActionAbortException extends \Exception
+{
+}